One floating-point type: the mechanical half, and the signatures that cannot convert
ROADMAP item 57. The game is to compute in float, with double left only
where a value provably needs it.

Two things had to happen before anything else, because nothing else can be
done a file at a time around them. Five virtual functions carry these
types - Object::onRender and reflectProjectile, Crossfade::render and the
two Upscaler warps - and a base whose signature changes while an override
does not still compiles, silently stops overriding, and the object stops
drawing. Those are changed across the tree in one pass. Sprites::sample
takes a pointer, and a pointer never converts, so it answers a Vec4f now.
Everything else crosses file boundaries by the conversion operator in
Vec<T, DIM>, which is what lets the tree compile at every step of this.

Then the mechanical pass over Blocks5/src: the vector aliases, the double
keyword, TinyXML's reader, which has a float spelling of its own, and the
floating literals, so that the arithmetic is float rather than double
narrowed at the end. The Vec2d, Vec3d and Vec4d aliases are gone from
typedefs.h with their last callers.

vec.h is deliberately untouched. Its Mat4 reproduces GLU's own arithmetic
in double, and rendering.md measures what that buys down to a right
angle's cosine of -4.4e-8.

The tree compiles: no errors from any of the 122 sources. What is left is
the project's own gate, an integer handed to a float, which this change
makes reach much further than it did, and the casts and comments that the
pass leaves behind.

+++ Blocks5/src/audiocapture.cpp
@@ -301,8 +301,8 @@ struct AudioCaptureImpl : public AudioRing
 	bool srcFloat;
 
 	// resampler state (capture thread only)
-	double resampleStep;
-	double resamplePos;
+	float resampleStep;
+	float resamplePos;
 	float prevLeft;
 	float prevRight;
 	bool havePrev;
@@ -323,8 +323,8 @@ AudioCaptureImpl::AudioCaptureImpl()
 	, srcBits(32)
 	, srcBlockAlign(8)
 	, srcFloat(true)
-	, resampleStep(1.0)
-	, resamplePos(0.0)
+	, resampleStep(1.0f)
+	, resamplePos(0.0f)
 	, prevLeft(0.0f)
 	, prevRight(0.0f)
 	, havePrev(false)
@@ -357,7 +357,7 @@ bool AudioCaptureImpl::setupSourceFormat(const WAVEFORMATEX* p_format)
 	if(!srcFloat && srcBits != 16 && srcBits != 24 && srcBits != 32) return false;
 	if(srcBlockAlign < srcChannels * (srcBits / 8)) return false;
 
-	resampleStep = (double)srcRate / (double)sampleRate;
+	resampleStep = (float)srcRate / (float)sampleRate;
 	return true;
 }
 
@@ -401,14 +401,14 @@ void AudioCaptureImpl::convertAndPush(const BYTE* p_data, int numFrames, bool si
 		{
 			prevLeft = left;
 			prevRight = right;
-			resamplePos = 0.0;
+			resamplePos = 0.0f;
 			havePrev = true;
 		}
 
 		// interpolate linearly between the previous and the current device sample.
 		// At an equal sample rate resampleStep is exactly 1.0 and every sample comes
 		// through unchanged.
-		while(resamplePos < 1.0)
+		while(resamplePos < 1.0f)
 		{
 			const float t = (float)resamplePos;
 			scratch[2 * numInScratch    ] = floatToShort(prevLeft  + (left  - prevLeft ) * t);
@@ -422,7 +422,7 @@ void AudioCaptureImpl::convertAndPush(const BYTE* p_data, int numFrames, bool si
 			}
 			resamplePos += resampleStep;
 		}
-		resamplePos -= 1.0;
+		resamplePos -= 1.0f;
 		prevLeft = left;
 		prevRight = right;
 	}
@@ -514,7 +514,7 @@ int AudioCaptureImpl::threadProc()
 				// throw away everything still lying around from last time
 				clearRing();
 				havePrev = false;
-				resamplePos = 0.0;
+				resamplePos = 0.0f;
 				samplesWritten = 0;
 				p_audioClient->Reset();
 				QueryPerformanceCounter(&captureStart);
